+++ src/components/main-screen/main-screen.jsx
@@ -6,7 +6,8 @@ import PropTypes from 'prop-types';
 import SortingOptions from '../sorting-options/sorting-options';
 import CitiesList from '../cities-list/cities-list';
 import PlacesList from '../places-list/places-list';
-import {CardName, sortCards, AuthorizationStatus, AppRoute} from '../../const';
+import {CardName, AuthorizationStatus, AppRoute} from '../../const';
+import {sortCards} from '../../utils';
 import Map from '../map/map';
 import {fetchHotelsList} from "../../store/api-actions";
 import SpinnerScreen from '../spinner-screen/spinner-screen';
